lvm: remove lvm2cmd code as it leaks memoty
